*** empty log message ***

+++ java/org/wewi/medimg/image/FloatData.java
@@ -1,5 +1,6 @@
 /**
- * Created on 22.10.2002 00:03:26
+ * FloatData.java
+ * Created on 13.06.2003
  *
  */
 package org.wewi.medimg.image;
@@ -10,30 +11,30 @@ import java.util.Arrays;
  * @author Franz Wilhelmstötter
  * @version 0.1
  */
-final class ByteData implements ImageData {
-    private byte[] data;
+final class FloatData implements ImageData {
+    private float[] data;
 
     /**
      * Constructor for ByteData.
      */
-    public ByteData(int size) {
-        data = new byte[size];
+    public FloatData(int size) {
+        data = new float[size];
     }
     
     public byte getByteValue(int pos) {
-        return data[pos];
+        return (byte)data[pos];
     }
     
     public short getShortValue(int pos) {
-        return data[pos];
+        return (short)data[pos];
     }
 
     public int getIntValue(int pos) {
-        return data[pos];
+        return (int)data[pos];
     }
     
     public long getLongValue(int pos) {
-        return data[pos];
+        return (long)data[pos];
     }
     
     public float getFloatValue(int pos) {
@@ -49,57 +50,56 @@ final class ByteData implements ImageData {
     }
 
     public void setValue(int pos, double value) {
-        data[pos] = (byte)value;
+        data[pos] = (float)value;
     }
 
     public void setValue(int pos, float value) {
-        data[pos] = (byte)value;
+        data[pos] = (float)value;
     }
 
     public void setValue(int pos, int value) {
-        data[pos] = (byte)value;
+        data[pos] = (float)value;
     }
 
     public void setValue(int pos, long value) {
-        data[pos] = (byte)value;
+        data[pos] = (float)value;
     }
 
     public void setValue(int pos, short value) {
-        data[pos] = (byte)value;
+        data[pos] = (float)value;
     }
 
     public void fill(byte value) {
-        Arrays.fill(data, value);
+        Arrays.fill(data, (float)value);
     }
 
     public void fill(double value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (float)value);
     }
 
     public void fill(float value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (float)value);
     }
 
     public void fill(int value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (float)value);
     }
 
     public void fill(long value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (float)value);
     }
 
     public void fill(short value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (float)value);
     }
 
 
     public void copy(ImageData target) {
-        if (!(target instanceof ByteData)) {
+        if (!(target instanceof FloatData)) {
             return;    
         }
         
-        ByteData t = (ByteData)target;
+        FloatData t = (FloatData)target;
         System.arraycopy(data, 0, t.data, 0, t.data.length);        
     }
-
 }
