+++ java/org/wewi/medimg/image/LongData.java
@@ -1,5 +1,6 @@
 /**
- * Created on 22.10.2002 00:03:26
+ * LongData.java
+ * Created on 13.06.2003
  *
  */
 package org.wewi.medimg.image;
@@ -10,26 +11,26 @@ import java.util.Arrays;
  * @author Franz Wilhelmstötter
  * @version 0.1
  */
-final class ByteData implements ImageData {
-    private byte[] data;
+final class LongData implements ImageData {
+    private long[] data;
 
     /**
      * Constructor for ByteData.
      */
-    public ByteData(int size) {
-        data = new byte[size];
+    public LongData(int size) {
+        data = new long[size];
     }
     
     public byte getByteValue(int pos) {
-        return data[pos];
+        return (byte)data[pos];
     }
     
     public short getShortValue(int pos) {
-        return data[pos];
+        return (short)data[pos];
     }
 
     public int getIntValue(int pos) {
-        return data[pos];
+        return (int)data[pos];
     }
     
     public long getLongValue(int pos) {
@@ -49,57 +50,56 @@ final class ByteData implements ImageData {
     }
 
     public void setValue(int pos, double value) {
-        data[pos] = (byte)value;
+        data[pos] = (long)value;
     }
 
     public void setValue(int pos, float value) {
-        data[pos] = (byte)value;
+        data[pos] = (long)value;
     }
 
     public void setValue(int pos, int value) {
-        data[pos] = (byte)value;
+        data[pos] = (long)value;
     }
 
     public void setValue(int pos, long value) {
-        data[pos] = (byte)value;
+        data[pos] = value;
     }
 
     public void setValue(int pos, short value) {
-        data[pos] = (byte)value;
+        data[pos] = (long)value;
     }
 
     public void fill(byte value) {
-        Arrays.fill(data, value);
+        Arrays.fill(data, (long)value);
     }
 
     public void fill(double value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (long)value);
     }
 
     public void fill(float value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (long)value);
     }
 
     public void fill(int value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (long)value);
     }
 
     public void fill(long value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (long)value);
     }
 
     public void fill(short value) {
-        Arrays.fill(data, (byte)value);
+        Arrays.fill(data, (long)value);
     }
 
 
     public void copy(ImageData target) {
-        if (!(target instanceof ByteData)) {
+        if (!(target instanceof LongData)) {
             return;    
         }
         
-        ByteData t = (ByteData)target;
+        LongData t = (LongData)target;
         System.arraycopy(data, 0, t.data, 0, t.data.length);        
     }
-
 }
